backend/requirements.txt
anyio==4.0.0
asgiref==3.6.0
astroid==2.15.6
async-timeout==4.0.3
attrs==23.1.0
certifi==2022.12.7
cfgv==3.3.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.0.1
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.6
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.5
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-extensions==0.7.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
filelock==3.9.0
flake8==6.0.0
flake8-django==1.3
gunicorn==20.1.0
h11==0.14.0
httptools==0.6.1
identify==2.5.18
idna==3.4
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
msgpack==1.0.7
nodeenv==1.7.0
openapi-codec==1.3.2
openpyxl==3.1.1
platformdirs==3.0.0
pre-commit==3.0.4
psycopg2==2.9.7
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycryptodome==3.18.0
pycryptodome-test-vectors==1.0.13
pyflakes==3.0.1
PyJWT==1.7.1
pylint==2.17.4
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
redis==5.0.1
referencing==0.32.0
requests==2.28.2
rpds-py==0.13.2
simplejson==3.18.3
sniffio==1.3.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.24.0.post1
uvloop==0.19.0
virtualenv==20.19.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.15.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
